// src/app/lib/types.ts
import firebase from 'firebase/compat/app';

// The base post data, excluding fields generated by the server.
export interface PostInput {
  userId: string;
  author: string;
  title: string;
  content: string;
}

// The complete Post type, including a generated ID and server timestamps.
export interface Post extends PostInput {
  id: string;
  imageUrl: string | null;
  createdAt: Date | firebase.firestore.FieldValue;
}

// The base comment data, excluding fields generated by the server.
export interface CommentInput {
  postId: string;
  author: string;
  content: string;
}

// The complete Comment type, including a generated ID and server timestamps.
export interface Comment extends CommentInput {
  id: string;
  postId: string;
  createdAt: Date | firebase.firestore.FieldValue;
}